R6 Class Representing a Cohort Manifest Item

The CohortManifestItem R6 class inherits from the ManifestItem R6 class.

Details

The CohortManifestItem is used to populate the CohortManifest object within a ManifestDb.

Super class

Barista::ManifestItem -> CohortManifestItem

Methods

Public methods

CohortManifestItem$new()

CohortManifestItem$clone()

Inherited methods

Barista::ManifestItem$loadItem()

Barista::ManifestItem$setFieldValue()

Barista::ManifestItem$setIdValue()

Method new()

Create a new CohortManifestItem object.

Usage

CohortManifestItem$new(
  provenanceId = -1,
  designMethod,
  name,
  relativeSqlPath = NA,
  relativeJsonPath = NA,
  relativeCaprPath = NA,
  relativeRPath = NA
)

Arguments

provenanceId

The id from the original design method system (e.g. Atlas id)

designMethod

The way in which the item was designed.

Choices for Concept Set: Atlas.

Choices for Cohort: Atlas, Subset, UnionTemplate, SqlTemplate, Capr.

name

The name of the item to add to the manifest.

relativeSqlPath

The relative path to the SQL file defining the manifest item.

Note: only 1 relative path can be defined for a manifest item.

relativeJsonPath

The relative path to the JSON file defining the manifest item.

Note: only 1 relative path can be defined for a manifest item.

relativeCaprPath

The relative path to the Capr R file defining the manifest item.

Note: only 1 relative path can be defined for a manifest item.

relativeRPath

The relative path to the R file defining the manifest item.

Note: only 1 relative path can be defined for a manifest item.

Method clone()

The objects of this class are cloneable with this method.

Usage

CohortManifestItem$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.