Collapse technical replicates

Collapse technical replicates in a DGEList object by mean or sum. Modified from collapseReplicates function from DESeq2 package to accept DGEList objects and to allow collapsing replicated by mean or sum.

Usage

collapseReplicates(
  object,
  groupby,
  run,
  renameCols = TRUE,
  show_reps = TRUE,
  by = "mean"
)

Arguments

object

DGEList object containing raw or normalised barcode counts with replicate grouping info in object metadata

groupby

a character vector of containing grouping information. Must be as long as the columns of object. Can pass a metadata column from object.

run

optional, the names of each unique column in object. if provided, a new column runsCollapsed will be added to the colData which pastes together the names of run

renameCols

boolean. whether to rename the columns of the returned object using the levels of the grouping factor

show_reps

boolean. print replicate column names to console?

by

collapse replicates by mean or sum

Value

Returns a dataframe of normalised counts for a sample

Examples

data(test.dge)
collapseReplicates(test.dge, groupby = test.dge$samples$group, by = "mean")
#> $`1_BM_High_dose`
#> [1] 41 42
#> 
#> $`1_SPL_High_dose`
#> [1] 43 44
#> 
#> $`10_BM_High_dose`
#> [1] 5 6
#> 
#> $`10_SPL_High_dose`
#> [1] 7 8
#> 
#> $`11_BM_Vehicle`
#> [1]  9 10
#> 
#> $`11_SPL_Vehicle`
#> [1] 11 12
#> 
#> $`12_BM_Vehicle`
#> [1] 13 14
#> 
#> $`12_SPL_Vehicle`
#> [1] 15 16
#> 
#> $`13_BM_Low_dose`
#> [1] 17 18
#> 
#> $`13_SPL_Low_dose`
#> [1] 19 20
#> 
#> $`14_BM_Low_dose`
#> [1] 21 22
#> 
#> $`14_SPL_Low_dose`
#> [1] 23 24
#> 
#> $`15_BM_Low_dose`
#> [1] 25 26
#> 
#> $`15_SPL_Low_dose`
#> [1] 27 28
#> 
#> $`16_BM_Low_dose`
#> [1] 29 30
#> 
#> $`16_SPL_Low_dose`
#> [1] 31 32
#> 
#> $`17_BM_High_dose`
#> [1] 33 34
#> 
#> $`17_SPL_High_dose`
#> [1] 35 36
#> 
#> $`18_BM_Vehicle`
#> [1] 37 38
#> 
#> $`18_SPL_Vehicle`
#> [1] 39 40
#> 
#> $`2_BM_High_dose`
#> [1] 45 46
#> 
#> $`2_SPL_High_dose`
#> [1] 47 48
#> 
#> $`3_BM_High_dose`
#> [1] 49 50
#> 
#> $`3_SPL_High_dose`
#> [1] 51 52
#> 
#> $`4_BM_Vehicle`
#> [1] 53 54
#> 
#> $`4_SPL_Vehicle`
#> [1] 55 56
#> 
#> $`5_BM_Low_dose`
#> [1] 57 58
#> 
#> $`5_SPL_Low_dose`
#> [1] 59 60
#> 
#> $`6_BM_High_dose`
#> [1] 61 62
#> 
#> $`6_SPL_High_dose`
#> [1] 63 64
#> 
#> $`7_BM_Vehicle`
#> [1] 65 66
#> 
#> $`7_SPL_Vehicle`
#> [1] 67 68
#> 
#> $`8_BM_Vehicle`
#> [1] 69 70
#> 
#> $`8_SPL_Vehicle`
#> [1] 71 72
#> 
#> $`9_BM_Low_dose`
#> [1] 73 74
#> 
#> $`9_SPL_Low_dose`
#> [1] 75 76
#> 
#> $T0
#> [1] 1 2
#> 
#> $T0b
#> [1] 3 4
#> 
#> An object of class "DGEList"
#> $samples
#>                   Sample Experiment            Group Strain Mouse Tissue
#> 1_BM_High_dose    1-BM-1       LM25   1_BM_High_dose C57BL6     1     BM
#> 1_SPL_High_dose   1-SP-1       LM25  1_SPL_High_dose C57BL6     1    SPL
#> 10_BM_High_dose  10-BM-1       LM25  10_BM_High_dose C57BL6    10     BM
#> 10_SPL_High_dose 10-SP-1       LM25 10_SPL_High_dose C57BL6    10    SPL
#> 11_BM_Vehicle    11-BM-1       LM25    11_BM_Vehicle C57BL6    11     BM
#>                  PCR_Replicate Treatment Days.post.transplant Library
#> 1_BM_High_dose               1 High_dose                   35     BFP
#> 1_SPL_High_dose              1 High_dose                   35     BFP
#> 10_BM_High_dose              1 High_dose                   41     BFP
#> 10_SPL_High_dose             1 High_dose                   41     BFP
#> 11_BM_Vehicle                1   Vehicle                   30     BFP
#>                             group lib.size norm.factors
#> 1_BM_High_dose     1_BM_High_dose  3447629            1
#> 1_SPL_High_dose   1_SPL_High_dose      246            1
#> 10_BM_High_dose   10_BM_High_dose  4114186            1
#> 10_SPL_High_dose 10_SPL_High_dose  4670972            1
#> 11_BM_Vehicle       11_BM_Vehicle  2907500            1
#> 33 more rows ...
#> 
#> $counts
#>                  Samples
#> Tags              1_BM_High_dose 1_SPL_High_dose 10_BM_High_dose
#>   BFP_Barcode_1                0               0               0
#>   BFP_Barcode_13               0               0               0
#>   BFP_Barcode_99             155               0               0
#>   BFP_Barcode_120              0               0               0
#>   BFP_Barcode_351              0               0               0
#>                  Samples
#> Tags              10_SPL_High_dose 11_BM_Vehicle 11_SPL_Vehicle 12_BM_Vehicle
#>   BFP_Barcode_1                  0             0              0             0
#>   BFP_Barcode_13                 0             0              2             0
#>   BFP_Barcode_99                 0             0              2             0
#>   BFP_Barcode_120                0             0              0             0
#>   BFP_Barcode_351                0             0              0             0
#>                  Samples
#> Tags              12_SPL_Vehicle 13_BM_Low_dose 13_SPL_Low_dose 14_BM_Low_dose
#>   BFP_Barcode_1                0              0               0              0
#>   BFP_Barcode_13               0              0               0              0
#>   BFP_Barcode_99               0              0               0              0
#>   BFP_Barcode_120              0              0               0              0
#>   BFP_Barcode_351              0              0               0              0
#>                  Samples
#> Tags              14_SPL_Low_dose 15_BM_Low_dose 15_SPL_Low_dose 16_BM_Low_dose
#>   BFP_Barcode_1                 0              0               0              0
#>   BFP_Barcode_13                0              0               0              0
#>   BFP_Barcode_99                0              0               0              0
#>   BFP_Barcode_120               0              0               0              0
#>   BFP_Barcode_351               0              0               0              0
#>                  Samples
#> Tags              16_SPL_Low_dose 17_BM_High_dose 17_SPL_High_dose
#>   BFP_Barcode_1                 0               0                0
#>   BFP_Barcode_13                0               0                0
#>   BFP_Barcode_99                0               0                0
#>   BFP_Barcode_120               0               0                0
#>   BFP_Barcode_351               0               0                0
#>                  Samples
#> Tags              18_BM_Vehicle 18_SPL_Vehicle 2_BM_High_dose 2_SPL_High_dose
#>   BFP_Barcode_1               0              0              0               0
#>   BFP_Barcode_13              0              0              0               0
#>   BFP_Barcode_99              0              0              0               0
#>   BFP_Barcode_120             0              0              0               0
#>   BFP_Barcode_351             0              0              0               0
#>                  Samples
#> Tags              3_BM_High_dose 3_SPL_High_dose 4_BM_Vehicle 4_SPL_Vehicle
#>   BFP_Barcode_1                0               0            0             0
#>   BFP_Barcode_13               0               0            0             0
#>   BFP_Barcode_99               0               0            0             0
#>   BFP_Barcode_120              0               0            0             0
#>   BFP_Barcode_351              0               0            0             0
#>                  Samples
#> Tags              5_BM_Low_dose 5_SPL_Low_dose 6_BM_High_dose 6_SPL_High_dose
#>   BFP_Barcode_1               0              0              0               0
#>   BFP_Barcode_13              0              0              0               0
#>   BFP_Barcode_99              0              0              0               0
#>   BFP_Barcode_120             0              0              0               0
#>   BFP_Barcode_351             0              0              0               0
#>                  Samples
#> Tags              7_BM_Vehicle 7_SPL_Vehicle 8_BM_Vehicle 8_SPL_Vehicle
#>   BFP_Barcode_1              0             0            0             0
#>   BFP_Barcode_13             0             0            0             0
#>   BFP_Barcode_99             0             0            0             0
#>   BFP_Barcode_120            0             0            0             0
#>   BFP_Barcode_351            0             0            0             0
#>                  Samples
#> Tags              9_BM_Low_dose 9_SPL_Low_dose   T0 T0b
#>   BFP_Barcode_1               0              0  127 352
#>   BFP_Barcode_13              0              0 1146 378
#>   BFP_Barcode_99              0              0 1354 989
#>   BFP_Barcode_120             0              0  234 236
#>   BFP_Barcode_351             0              0    0  61
#> 1634 more rows ...
#>