Get dominant barcodes per sample

Takes a DGEList or dataframe of barcode counts, computes the percentage abundance of each barcode within each sample then returns a list of barcodes meeting a threshold abundance per sample

Usage

getDominantBarcodes(counts.obj, pct.thresh = 5)

Arguments

counts.obj

DGEList object or dataframe containing raw / normalised counts of barcodes

pct.thresh

percentage abundance threshold to call dominant barcodes

Value

Returns a named list containing vectors of dominant barcodes per sample

Details

getDominantBarcodes

get names of barcodes over a given percentage frequency per sample

Examples

data(test.dge)
getDominantBarcodes(test.dge, pct.thresh = 5)
#> $`T0-1`
#> [1] "BFP_Barcode_245016" "BFP_Barcode_121651"
#> 
#> $`T0-2`
#> [1] "BFP_Barcode_245016" "BFP_Barcode_121651"
#> 
#> $`T0b-1`
#> [1] "BFP_Barcode_121651" "BFP_Barcode_245016"
#> 
#> $`T0b-2`
#> [1] "BFP_Barcode_245016" "BFP_Barcode_121651"
#> 
#> $`10-BM-1`
#> [1] "BFP_Barcode_8419"   "BFP_Barcode_124796"
#> 
#> $`10-BM-2`
#> [1] "BFP_Barcode_8419"   "BFP_Barcode_124796"
#> 
#> $`10-SP-1`
#> [1] "BFP_Barcode_8419"   "BFP_Barcode_124796" "BFP_Barcode_102160"
#> 
#> $`10-SP-2`
#> [1] "BFP_Barcode_8419"   "BFP_Barcode_124796" "BFP_Barcode_102160"
#> 
#> $`11-BM-1`
#> [1] "BFP_Barcode_53234"  "BFP_Barcode_205581" "BFP_Barcode_90135" 
#> [4] "BFP_Barcode_172626" "BFP_Barcode_58978" 
#> 
#> $`11-BM-2`
#> [1] "BFP_Barcode_53234"  "BFP_Barcode_205581" "BFP_Barcode_90135" 
#> [4] "BFP_Barcode_172626" "BFP_Barcode_58978" 
#> 
#> $`11-SP-1`
#> [1] "BFP_Barcode_53234"  "BFP_Barcode_205581" "BFP_Barcode_90135" 
#> [4] "BFP_Barcode_159570" "BFP_Barcode_172626" "BFP_Barcode_257382"
#> [7] "BFP_Barcode_79755" 
#> 
#> $`11-SP-2`
#> [1] "BFP_Barcode_53234"  "BFP_Barcode_205581" "BFP_Barcode_90135" 
#> [4] "BFP_Barcode_159570" "BFP_Barcode_172626" "BFP_Barcode_257382"
#> [7] "BFP_Barcode_79755" 
#> 
#> $`12-BM-1`
#> [1] "BFP_Barcode_389078" "BFP_Barcode_159570" "BFP_Barcode_135438"
#> [4] "BFP_Barcode_500780" "BFP_Barcode_79755" 
#> 
#> $`12-BM-2`
#> [1] "BFP_Barcode_389078" "BFP_Barcode_159570" "BFP_Barcode_135438"
#> [4] "BFP_Barcode_500780" "BFP_Barcode_79755" 
#> 
#> $`12-SP-1`
#> [1] "BFP_Barcode_389078" "BFP_Barcode_159570" "BFP_Barcode_49629" 
#> [4] "BFP_Barcode_500780" "BFP_Barcode_31610"  "BFP_Barcode_79755" 
#> 
#> $`12-SP-2`
#> [1] "BFP_Barcode_389078" "BFP_Barcode_159570" "BFP_Barcode_49629" 
#> [4] "BFP_Barcode_500780" "BFP_Barcode_31610" 
#> 
#> $`13-BM-1`
#> [1] "BFP_Barcode_8419"   "BFP_Barcode_124796" "BFP_Barcode_102160"
#> [4] "BFP_Barcode_257382"
#> 
#> $`13-BM-2`
#> [1] "BFP_Barcode_8419"   "BFP_Barcode_124796" "BFP_Barcode_102160"
#> [4] "BFP_Barcode_257382"
#> 
#> $`13-SP-1`
#> [1] "BFP_Barcode_8419"   "BFP_Barcode_124796" "BFP_Barcode_257382"
#> [4] "BFP_Barcode_102160"
#> 
#> $`13-SP-2`
#> [1] "BFP_Barcode_8419"   "BFP_Barcode_124796" "BFP_Barcode_257382"
#> [4] "BFP_Barcode_102160"
#> 
#> $`14-BM-1`
#> [1] "BFP_Barcode_90135"  "BFP_Barcode_388103" "BFP_Barcode_31610" 
#> 
#> $`14-BM-2`
#> [1] "BFP_Barcode_90135"  "BFP_Barcode_388103" "BFP_Barcode_31610" 
#> 
#> $`14-SP-1`
#> [1] "BFP_Barcode_388103" "BFP_Barcode_90135"  "BFP_Barcode_31610" 
#> [4] "BFP_Barcode_292962"
#> 
#> $`14-SP-2`
#> [1] "BFP_Barcode_388103" "BFP_Barcode_90135"  "BFP_Barcode_31610" 
#> [4] "BFP_Barcode_292962"
#> 
#> $`15-BM-1`
#> [1] "BFP_Barcode_1478"   "BFP_Barcode_94579"  "BFP_Barcode_124796"
#> [4] "BFP_Barcode_8419"  
#> 
#> $`15-BM-2`
#> [1] "BFP_Barcode_1478"   "BFP_Barcode_94579"  "BFP_Barcode_124796"
#> [4] "BFP_Barcode_8419"  
#> 
#> $`15-SP-1`
#> [1] "BFP_Barcode_1478"   "BFP_Barcode_124796" "BFP_Barcode_8419"  
#> [4] "BFP_Barcode_94579"  "BFP_Barcode_400391"
#> 
#> $`15-SP-2`
#> [1] "BFP_Barcode_1478"   "BFP_Barcode_124796" "BFP_Barcode_8419"  
#> [4] "BFP_Barcode_94579"  "BFP_Barcode_400391"
#> 
#> $`16-BM-1`
#> [1] "BFP_Barcode_53234"  "BFP_Barcode_79755"  "BFP_Barcode_140320"
#> [4] "BFP_Barcode_31610"  "BFP_Barcode_248880" "BFP_Barcode_1478"  
#> 
#> $`16-BM-2`
#> [1] "BFP_Barcode_53234"  "BFP_Barcode_79755"  "BFP_Barcode_140320"
#> [4] "BFP_Barcode_31610"  "BFP_Barcode_248880" "BFP_Barcode_1478"  
#> 
#> $`16-SP-1`
#> [1] "BFP_Barcode_248880" "BFP_Barcode_79755"  "BFP_Barcode_1478"  
#> [4] "BFP_Barcode_257382" "BFP_Barcode_142112" "BFP_Barcode_53234" 
#> [7] "BFP_Barcode_94511"  "BFP_Barcode_31610" 
#> 
#> $`16-SP-2`
#> [1] "BFP_Barcode_248880" "BFP_Barcode_79755"  "BFP_Barcode_1478"  
#> [4] "BFP_Barcode_257382" "BFP_Barcode_142112" "BFP_Barcode_53234" 
#> [7] "BFP_Barcode_94511"  "BFP_Barcode_31610"  "BFP_Barcode_142024"
#> 
#> $`17-BM-1`
#> [1] "BFP_Barcode_183691" "BFP_Barcode_90135"  "BFP_Barcode_400391"
#> 
#> $`17-BM-2`
#> [1] "BFP_Barcode_183691" "BFP_Barcode_90135"  "BFP_Barcode_400391"
#> 
#> $`17-SP-1`
#> [1] "BFP_Barcode_183691" "BFP_Barcode_400391" "BFP_Barcode_90135" 
#> 
#> $`17-SP-2`
#> [1] "BFP_Barcode_183691" "BFP_Barcode_400391" "BFP_Barcode_90135" 
#> 
#> $`18-BM-1`
#> [1] "BFP_Barcode_49397"  "BFP_Barcode_91412"  "BFP_Barcode_159570"
#> [4] "BFP_Barcode_62602"  "BFP_Barcode_93485"  "BFP_Barcode_70225" 
#> 
#> $`18-BM-2`
#> [1] "BFP_Barcode_49397"  "BFP_Barcode_91412"  "BFP_Barcode_159570"
#> [4] "BFP_Barcode_62602"  "BFP_Barcode_70225"  "BFP_Barcode_93485" 
#> 
#> $`18-SP-1`
#> [1] "BFP_Barcode_159570" "BFP_Barcode_49397"  "BFP_Barcode_4469"  
#> [4] "BFP_Barcode_91412" 
#> 
#> $`18-SP-2`
#> [1] "BFP_Barcode_159570" "BFP_Barcode_49397"  "BFP_Barcode_4469"  
#> [4] "BFP_Barcode_91412"  "BFP_Barcode_79755" 
#> 
#> $`1-BM-1`
#> [1] "BFP_Barcode_79755"  "BFP_Barcode_102160" "BFP_Barcode_59493" 
#> [4] "BFP_Barcode_23361"  "BFP_Barcode_53234"  "BFP_Barcode_400391"
#> [7] "BFP_Barcode_142112"
#> 
#> $`1-BM-2`
#> [1] "BFP_Barcode_79755"  "BFP_Barcode_102160" "BFP_Barcode_59493" 
#> [4] "BFP_Barcode_23361"  "BFP_Barcode_53234" 
#> 
#> $`1-SP-1`
#> [1] "BFP_Barcode_79755"  "BFP_Barcode_59493"  "BFP_Barcode_102160"
#> [4] "BFP_Barcode_23361" 
#> 
#> $`1-SP-2`
#> [1] "BFP_Barcode_79755"  "BFP_Barcode_59493"  "BFP_Barcode_102160"
#> [4] "BFP_Barcode_23361"  "BFP_Barcode_94511"  "BFP_Barcode_400391"
#> 
#> $`2-BM-1`
#> [1] "BFP_Barcode_248880" "BFP_Barcode_65104"  "BFP_Barcode_9141"  
#> [4] "BFP_Barcode_1478"   "BFP_Barcode_31610"  "BFP_Barcode_79755" 
#> 
#> $`2-BM-2`
#> [1] "BFP_Barcode_248880" "BFP_Barcode_65104"  "BFP_Barcode_9141"  
#> [4] "BFP_Barcode_1478"   "BFP_Barcode_31610"  "BFP_Barcode_79755" 
#> 
#> $`2-SP-1`
#> [1] "BFP_Barcode_248880" "BFP_Barcode_65104"  "BFP_Barcode_9141"  
#> [4] "BFP_Barcode_79755"  "BFP_Barcode_1478"   "BFP_Barcode_31610" 
#> 
#> $`2-SP-2`
#> [1] "BFP_Barcode_248880" "BFP_Barcode_65104"  "BFP_Barcode_9141"  
#> [4] "BFP_Barcode_79755"  "BFP_Barcode_1478"  
#> 
#> $`3-BM-1`
#> [1] "BFP_Barcode_79755"  "BFP_Barcode_142024" "BFP_Barcode_102160"
#> 
#> $`3-BM-2`
#> [1] "BFP_Barcode_79755"  "BFP_Barcode_142024" "BFP_Barcode_102160"
#> [4] "BFP_Barcode_183691"
#> 
#> $`3-SP-1`
#> [1] "BFP_Barcode_79755"  "BFP_Barcode_142024" "BFP_Barcode_183691"
#> [4] "BFP_Barcode_102160" "BFP_Barcode_59493"  "BFP_Barcode_23361" 
#> 
#> $`3-SP-2`
#> [1] "BFP_Barcode_79755"  "BFP_Barcode_142024" "BFP_Barcode_183691"
#> [4] "BFP_Barcode_102160" "BFP_Barcode_59493"  "BFP_Barcode_23361" 
#> 
#> $`4-BM-1`
#> [1] "BFP_Barcode_53234"  "BFP_Barcode_58978"  "BFP_Barcode_62602" 
#> [4] "BFP_Barcode_68847"  "BFP_Barcode_140320" "BFP_Barcode_79755" 
#> 
#> $`4-BM-2`
#> [1] "BFP_Barcode_53234"  "BFP_Barcode_58978"  "BFP_Barcode_62602" 
#> [4] "BFP_Barcode_68847"  "BFP_Barcode_140320" "BFP_Barcode_79755" 
#> [7] "BFP_Barcode_327422"
#> 
#> $`4-SP-1`
#> [1] "BFP_Barcode_53234"  "BFP_Barcode_327422" "BFP_Barcode_79755" 
#> [4] "BFP_Barcode_58978"  "BFP_Barcode_81579"  "BFP_Barcode_159570"
#> 
#> $`4-SP-2`
#> [1] "BFP_Barcode_53234"  "BFP_Barcode_327422" "BFP_Barcode_79755" 
#> [4] "BFP_Barcode_58978"  "BFP_Barcode_81579" 
#> 
#> $`5-BM-1`
#> [1] "BFP_Barcode_561471" "BFP_Barcode_118640" "BFP_Barcode_59493" 
#> [4] "BFP_Barcode_1607"  
#> 
#> $`5-BM-2`
#> [1] "BFP_Barcode_561471" "BFP_Barcode_118640" "BFP_Barcode_59493" 
#> [4] "BFP_Barcode_1607"   "BFP_Barcode_31610" 
#> 
#> $`5-SP-1`
#> [1] "BFP_Barcode_561471" "BFP_Barcode_59493"  "BFP_Barcode_118640"
#> [4] "BFP_Barcode_23361"  "BFP_Barcode_1607"   "BFP_Barcode_31610" 
#> 
#> $`5-SP-2`
#> [1] "BFP_Barcode_561471" "BFP_Barcode_59493"  "BFP_Barcode_23361" 
#> [4] "BFP_Barcode_118640" "BFP_Barcode_1607"   "BFP_Barcode_31610" 
#> 
#> $`6-BM-1`
#> [1] "BFP_Barcode_79755"  "BFP_Barcode_8419"   "BFP_Barcode_124796"
#> [4] "BFP_Barcode_4564"  
#> 
#> $`6-BM-2`
#> [1] "BFP_Barcode_79755"  "BFP_Barcode_8419"   "BFP_Barcode_124796"
#> [4] "BFP_Barcode_4564"  
#> 
#> $`6-SP-1`
#> [1] "BFP_Barcode_8419"   "BFP_Barcode_124796" "BFP_Barcode_79755" 
#> 
#> $`6-SP-2`
#> [1] "BFP_Barcode_8419"   "BFP_Barcode_124796" "BFP_Barcode_79755" 
#> [4] "BFP_Barcode_400391"
#> 
#> $`7-BM-1`
#> [1] "BFP_Barcode_53234"  "BFP_Barcode_54442"  "BFP_Barcode_142024"
#> [4] "BFP_Barcode_140320" "BFP_Barcode_93485" 
#> 
#> $`7-BM-2`
#> [1] "BFP_Barcode_53234"  "BFP_Barcode_54442"  "BFP_Barcode_142024"
#> [4] "BFP_Barcode_140320" "BFP_Barcode_93485" 
#> 
#> $`7-SP-1`
#> [1] "BFP_Barcode_142024" "BFP_Barcode_54442"  "BFP_Barcode_53234" 
#> [4] "BFP_Barcode_93485"  "BFP_Barcode_159570"
#> 
#> $`7-SP-2`
#> [1] "BFP_Barcode_142024" "BFP_Barcode_54442"  "BFP_Barcode_53234" 
#> [4] "BFP_Barcode_93485"  "BFP_Barcode_159570"
#> 
#> $`8-BM-1`
#> [1] "BFP_Barcode_142024" "BFP_Barcode_90135"  "BFP_Barcode_58978" 
#> [4] "BFP_Barcode_93485" 
#> 
#> $`8-BM-2`
#> [1] "BFP_Barcode_142024" "BFP_Barcode_90135"  "BFP_Barcode_58978" 
#> [4] "BFP_Barcode_93485" 
#> 
#> $`8-SP-1`
#> [1] "BFP_Barcode_142024" "BFP_Barcode_90135"  "BFP_Barcode_4173"  
#> [4] "BFP_Barcode_58978"  "BFP_Barcode_31610" 
#> 
#> $`8-SP-2`
#> [1] "BFP_Barcode_142024" "BFP_Barcode_90135"  "BFP_Barcode_4173"  
#> [4] "BFP_Barcode_58978" 
#> 
#> $`9-BM-1`
#> [1] "BFP_Barcode_79755"  "BFP_Barcode_4198"   "BFP_Barcode_81579" 
#> [4] "BFP_Barcode_560839" "BFP_Barcode_118640" "BFP_Barcode_345889"
#> [7] "BFP_Barcode_400391"
#> 
#> $`9-BM-2`
#> [1] "BFP_Barcode_79755"  "BFP_Barcode_4198"   "BFP_Barcode_81579" 
#> [4] "BFP_Barcode_118640" "BFP_Barcode_560839" "BFP_Barcode_345889"
#> [7] "BFP_Barcode_400391"
#> 
#> $`9-SP-1`
#> [1] "BFP_Barcode_79755"  "BFP_Barcode_142024" "BFP_Barcode_400391"
#> [4] "BFP_Barcode_81579"  "BFP_Barcode_19146"  "BFP_Barcode_4198"  
#> [7] "BFP_Barcode_118640"
#> 
#> $`9-SP-2`
#> [1] "BFP_Barcode_79755"  "BFP_Barcode_142024" "BFP_Barcode_400391"
#> [4] "BFP_Barcode_81579"  "BFP_Barcode_19146"  "BFP_Barcode_118640"
#> [7] "BFP_Barcode_4198"  
#>